+++ llamafile/sgemm.h
@@ -7,24 +7,24 @@ extern "C" {
 struct ggml_tensor;
 struct ggml_compute_params;
 
-bool llamafile_sgemm_unsupported(int, int, int, const void *, int, const void *, int, void *, int,
-                                 int, int, int, int, int, int);
-bool llamafile_sgemm_amd_avx(int, int, int, const void *, int, const void *, int, void *, int, int,
-                             int, int, int, int, int);
-bool llamafile_sgemm_amd_fma(int, int, int, const void *, int, const void *, int, void *, int, int,
-                             int, int, int, int, int);
-bool llamafile_sgemm_amd_avx2(int, int, int, const void *, int, const void *, int, void *, int, int,
-                              int, int, int, int, int);
-bool llamafile_sgemm_amd_avxvnni(int, int, int, const void *, int, const void *, int, void *, int,
-                                 int, int, int, int, int, int);
-bool llamafile_sgemm_amd_avx512f(int, int, int, const void *, int, const void *, int, void *, int,
-                                 int, int, int, int, int, int);
-bool llamafile_sgemm_amd_zen4(int, int, int, const void *, int, const void *, int, void *, int, int,
-                              int, int, int, int, int);
-bool llamafile_sgemm_arm80(int, int, int, const void *, int, const void *, int, void *, int, int,
-                           int, int, int, int, int);
-bool llamafile_sgemm_arm82(int, int, int, const void *, int, const void *, int, void *, int, int,
-                           int, int, int, int, int);
+bool llamafile_sgemm_unsupported(long, long, long, const void *, long, const void *, long, void *,
+                                 long, int, int, int, int, int, int);
+bool llamafile_sgemm_amd_avx(long, long, long, const void *, long, const void *, long, void *, long,
+                             int, int, int, int, int, int);
+bool llamafile_sgemm_amd_fma(long, long, long, const void *, long, const void *, long, void *, long,
+                             int, int, int, int, int, int);
+bool llamafile_sgemm_amd_avx2(long, long, long, const void *, long, const void *, long, void *,
+                              long, int, int, int, int, int, int);
+bool llamafile_sgemm_amd_avxvnni(long, long, long, const void *, long, const void *, long, void *,
+                                 long, int, int, int, int, int, int);
+bool llamafile_sgemm_amd_avx512f(long, long, long, const void *, long, const void *, long, void *,
+                                 long, int, int, int, int, int, int);
+bool llamafile_sgemm_amd_zen4(long, long, long, const void *, long, const void *, long, void *,
+                              long, int, int, int, int, int, int);
+bool llamafile_sgemm_arm80(long, long, long, const void *, long, const void *, long, void *, long,
+                           int, int, int, int, int, int);
+bool llamafile_sgemm_arm82(long, long, long, const void *, long, const void *, long, void *, long,
+                           int, int, int, int, int, int);
 
 bool llamafile_mixmul_unsupported(const struct ggml_compute_params *, const struct ggml_tensor *,
                                   const struct ggml_tensor *, const struct ggml_tensor *,
